create separate module for command and client, fix unit tests
